astro_server/requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.11
PyJWT~=2.10.1
passlib~=1.7.4
dotenv~=0.9.9
python-dotenv~=1.0.1
asyncpg~=0.30.0
pydantic~=2.10.6
bcrypt~=4.3.0